second stage. added network to movie details screen. added genres to movies screen/// fixes after review #2
