+++ app/src/main/java/com/dreamisi/moviebase/data/jsonmodels/JsonActor.kt
@@ -1,4 +1,4 @@
-package com.dreamisi.moviebase.data.JsonModels
+package com.dreamisi.moviebase.data.jsonmodels
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
