+++ app/src/main/java/com/dreamisi/moviebase/moviedetails/MovieDetailsViewModelFactory.kt
@@ -1,11 +1,11 @@
-package com.dreamisi.moviebase.movieList
+package com.dreamisi.moviebase.moviedetails
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.CreationExtras
 import com.dreamisi.moviebase.MovieBaseApplication
 
-class MoviesListViewModelFactory : ViewModelProvider.Factory {
+class MovieDetailsViewModelFactory(private val movieID: Int) : ViewModelProvider.Factory {
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(
         modelClass: Class<T>,
@@ -14,8 +14,9 @@ class MoviesListViewModelFactory : ViewModelProvider.Factory {
         val application =
             checkNotNull(extras[ViewModelProvider.AndroidViewModelFactory.APPLICATION_KEY])
 
-        return MoviesListViewModel(
-            (application as MovieBaseApplication).provideRepository()
+        return MovieDetailsViewModel(
+            (application as MovieBaseApplication).provideRepository(),
+            movieID
         ) as T
 
     }
