+++ app/src/main/java/com/dreamisi/moviebase/movielist/MoviesListViewModelFactory.kt
@@ -1,11 +1,11 @@
-package com.dreamisi.moviebase.movieDetails
+package com.dreamisi.moviebase.movielist
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.CreationExtras
 import com.dreamisi.moviebase.MovieBaseApplication
 
-class MovieDetailsViewModelFactory(private val movieID: Int) : ViewModelProvider.Factory {
+class MoviesListViewModelFactory : ViewModelProvider.Factory {
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(
         modelClass: Class<T>,
@@ -14,9 +14,8 @@ class MovieDetailsViewModelFactory(private val movieID: Int) : ViewModelProvider
         val application =
             checkNotNull(extras[ViewModelProvider.AndroidViewModelFactory.APPLICATION_KEY])
 
-        return MovieDetailsViewModel(
-            (application as MovieBaseApplication).provideRepository(),
-            movieID
+        return MoviesListViewModel(
+            (application as MovieBaseApplication).provideRepository()
         ) as T
 
     }
